API-16: improve policy into controller

+++ app/Http/Controllers/Question/DeleteController.php
@@ -2,13 +2,18 @@
 
 namespace App\Http\Controllers\Question;
 
-use App\Http\Requests\Question\DeleteRequest;
+use App\Http\Controllers\Controller;
 use App\Models\Question;
+use Illuminate\Support\Facades\Gate;
 
-class DeleteController
+class DeleteController extends Controller
 {
-    public function __invoke(Question $question, DeleteRequest $deleteRequest)
+    public function __invoke(Question $question)
     {
+        if (!Gate::allows('forceDelete', $question)) {
+            abort(403);
+        }
+
         $question->forceDelete();
 
         return response()->noContent();
